installer/launcher/requirements.txt
pyinstaller
pillow
qrcode[pil]
pyperclip
cryptography
